requirements.txt
# Please follow suitable version specifiers, as outlined in https://www.python.org/dev/peps/pep-0440/#version-specifiers
# For Django versioning we use ~= (e.g. Django~=4.2.0) which will keep the minor version up to date (e.g. the latest version prior to 4.3.0)

Django~=4.2.0
Pillow~=9.5.0
django-recaptcha~=3.0.0
django-embed-video~=1.4.9
django-ckeditor~=6.5.1
django-debug-toolbar~=4.1.0
django-admin-list-filter-dropdown~=1.0.3
coverage
flake8